+++ src/backend/winit/input.rs
@@ -7,7 +7,7 @@ use winit::{
 
 use crate::backend::input::{
     self, AbsolutePositionEvent, Axis, AxisRelativeDirection, AxisSource, ButtonState, Device,
-    DeviceCapability, Event, InputBackend, KeyState, KeyboardKeyEvent, Keycode, PointerAxisEvent,
+    DeviceCapability, Event, InputBackend, InputTime, KeyState, KeyboardKeyEvent, Keycode, PointerAxisEvent,
     PointerButtonEvent, PointerMotionAbsoluteEvent, TouchCancelEvent, TouchDownEvent, TouchEvent,
     TouchMotionEvent, TouchSlot, TouchUpEvent, UnusedEvent,
 };
@@ -48,14 +48,14 @@ impl Device for WinitVirtualDevice {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`KeyboardKeyEvent`]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitKeyboardInputEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) key: u32,
     pub(crate) count: u32,
     pub(crate) state: ElementState,
 }
 
 impl Event<WinitInput> for WinitKeyboardInputEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -81,13 +81,13 @@ impl KeyboardKeyEvent<WinitInput> for WinitKeyboardInputEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`PointerMotionAbsoluteEvent`]
 #[derive(Debug, Clone)]
 pub struct WinitMouseMovedEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) position: RelativePosition,
     pub(crate) global_position: PhysicalPosition<f64>,
 }
 
 impl Event<WinitInput> for WinitMouseMovedEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -118,12 +118,12 @@ impl AbsolutePositionEvent<WinitInput> for WinitMouseMovedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`PointerAxisEvent`]
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct WinitMouseWheelEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) delta: MouseScrollDelta,
 }
 
 impl Event<WinitInput> for WinitMouseWheelEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -137,6 +137,7 @@ impl PointerAxisEvent<WinitInput> for WinitMouseWheelEvent {
         match self.delta {
             MouseScrollDelta::LineDelta(_, _) => AxisSource::Wheel,
             MouseScrollDelta::PixelDelta(_) => AxisSource::Continuous,
+            _ => AxisSource::Continuous,
         }
     }
 
@@ -144,7 +145,7 @@ impl PointerAxisEvent<WinitInput> for WinitMouseWheelEvent {
         match (axis, self.delta) {
             (Axis::Horizontal, MouseScrollDelta::PixelDelta(delta)) => Some(-delta.x),
             (Axis::Vertical, MouseScrollDelta::PixelDelta(delta)) => Some(-delta.y),
-            (_, MouseScrollDelta::LineDelta(_, _)) => None,
+            _ => None,
         }
     }
 
@@ -153,7 +154,7 @@ impl PointerAxisEvent<WinitInput> for WinitMouseWheelEvent {
         match (axis, self.delta) {
             (Axis::Horizontal, MouseScrollDelta::LineDelta(x, _)) => Some(-x as f64 * 120.),
             (Axis::Vertical, MouseScrollDelta::LineDelta(_, y)) => Some(-y as f64 * 120.),
-            (_, MouseScrollDelta::PixelDelta(_)) => None,
+            _ => None,
         }
     }
 
@@ -166,14 +167,14 @@ impl PointerAxisEvent<WinitInput> for WinitMouseWheelEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`PointerButtonEvent`]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitMouseInputEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) button: WinitMouseButton,
     pub(crate) state: ElementState,
     pub(crate) is_x11: bool,
 }
 
 impl Event<WinitInput> for WinitMouseInputEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -208,14 +209,14 @@ impl PointerButtonEvent<WinitInput> for WinitMouseInputEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`TouchDownEvent`]
 #[derive(Debug, Clone)]
 pub struct WinitTouchStartedEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) position: RelativePosition,
     pub(crate) global_position: PhysicalPosition<f64>,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchStartedEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -253,14 +254,14 @@ impl AbsolutePositionEvent<WinitInput> for WinitTouchStartedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`TouchMotionEvent`]
 #[derive(Debug, Clone)]
 pub struct WinitTouchMovedEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) position: RelativePosition,
     pub(crate) global_position: PhysicalPosition<f64>,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchMovedEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -298,12 +299,12 @@ impl AbsolutePositionEvent<WinitInput> for WinitTouchMovedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a `TouchUpEvent`
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitTouchEndedEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchEndedEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }
 
@@ -323,12 +324,12 @@ impl TouchEvent<WinitInput> for WinitTouchEndedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`TouchCancelEvent`]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitTouchCancelledEvent {
-    pub(crate) time: u64,
+    pub(crate) time: InputTime,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchCancelledEvent {
-    fn time(&self) -> u64 {
+    fn time(&self) -> InputTime {
         self.time
     }